% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.iotjobsdataplane_operations.R
\name{get_pending_job_executions}
\alias{get_pending_job_executions}
\title{Gets the list of all jobs for a thing that are not in a terminal status}
\usage{
get_pending_job_executions(thingName)
}
\arguments{
\item{thingName}{The name of the thing that is executing the job.}
}
\description{
Gets the list of all jobs for a thing that are not in a terminal status.
}
\examples{

}
