% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.iotjobsdataplane_operations.R
\name{update_job_execution}
\alias{update_job_execution}
\title{Updates the status of a job execution}
\usage{
update_job_execution(
  jobId = "string",
  thingName = "string",
  status = "QUEUED"|"IN_PROGRESS"|"SUCCEEDED"|"FAILED"|"TIMED_OUT"|"REJECTED"|"REMOVED"|"CANCELED",
  statusDetails = list(
    "string"
  ),
  stepTimeoutInMinutes = 123,
  expectedVersion = 123,
  includeJobExecutionState = TRUE|FALSE,
  includeJobDocument = TRUE|FALSE,
  executionNumber = 123
)
}
\arguments{
\item{jobId}{[required] The unique identifier assigned to this job when it was created.}

\item{thingName}{[required] The name of the thing associated with the device.}

\item{status}{[required] The new status for the job execution (IN\_PROGRESS, FAILED, SUCCESS, or REJECTED). This must be specified on every update.}

\item{statusDetails}{Optional. A collection of name/value pairs that describe the status of the job execution. If not specified, the statusDetails are unchanged.}

\item{stepTimeoutInMinutes}{Specifies the amount of time this device has to finish execution of this job. If the job execution status is not set to a terminal state before this timer expires, or before the timer is reset (by again calling \code{UpdateJobExecution}, setting the status to \code{IN_PROGRESS} and specifying a new timeout value in this field) the job execution status will be automatically set to \code{TIMED_OUT}. Note that setting or resetting this timeout has no effect on that job execution timeout which may have been specified when the job was created (\code{CreateJob} using field \code{timeoutConfig}).}

\item{expectedVersion}{Optional. The expected current version of the job execution. Each time you update the job execution, its version is incremented. If the version of the job execution stored in Jobs does not match, the update is rejected with a VersionMismatch error, and an ErrorResponse that contains the current job execution status data is returned. (This makes it unnecessary to perform a separate DescribeJobExecution request in order to obtain the job execution status data.)}

\item{includeJobExecutionState}{Optional. When included and set to true, the response contains the JobExecutionState data. The default is false.}

\item{includeJobDocument}{Optional. When set to true, the response contains the job document. The default is false.}

\item{executionNumber}{Optional. A number that identifies a particular job execution on a particular device.}
}
\description{
Updates the status of a job execution.
}
\examples{

}
